Category Sorting
Added in mobileDevice category, and narrowing down specifications with categorization. Remembered that can use 'and' and 'not in' and that saved me a ton of headache. Lens and camera body seem to be fine as well as printers, but the issue is fixed lens (possibly mobileDevice) and misc. Can't figure out why some options are copying over to misc and why the sums aren't adding up to 4144... Tbd